Backend/requirements.txt
Django>=4.2,<5.1
Pillow>=10.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3
python-dotenv>=1.0
openai>=1.40
# PDF: PyMuPDF first (better layout text); pypdf fallback in code
pymupdf>=1.24.0
pypdf>=4.0
python-docx>=1.1
redis>=5.0,<6
gunicorn>=22.0
whitenoise>=6.6
dj-database-url>=2.2
psycopg2-binary>=2.9
django-cloudinary-storage>=0.3
cloudinary>=1.40
django-storages[s3]>=1.14
boto3>=1.34
celery[redis]>=5.3,<6
imageio-ffmpeg>=0.5.1
requests>=2.31
stripe>=15.0
stripe>=15.0
